Using mountable routes in express@^3, fix #9
