+++ index.js
@@ -1,52 +1,8 @@
-
-// See http://expressjs.com/guide.html
-
 var express = require('express');
-var app = express();
+var router = new express.Router();
 
-var mime = require('mime');
-var fs = require('fs');
-var $rdf = require('rdflib.js')
-
-// Should be command line params:
-
-var uriBase = '/test/' // @@
-var fileBase = '/devel/github.com/linkeddata/node-ldp-httpd/test/'; //@@
-
-var uriFilter = /\/test\/.*/
-
-var PATCH = $rdf.Namespace('http://www.w3.org/ns/pim/patch#');
-
-var uriToFilename = function(uri) {
-    if (uri.slice(0, uriBase.length) !== uriBase) {
-        throw "URI not starting with base: " + uriBase;
-    }
-    var filename = fileBase + uri.slice(uriBase.length);
-    console.log(' -- filename ' +filename);
-    return filename    
-};
-
-
-// See https://github.com/stream-utils/raw-body
-var getRawBody = require('raw-body')
-//var typer      = require('media-typer')
-app.use(function (req, res, next) {
-    getRawBody(req, {
-        length: req.headers['content-length'],
-        limit: '1mb',
-        encoding: 'utf-8' // typer.parse(req.headers['content-type']).parameters.charset
-    }, function (err, string) {
-    if (err) {
-        return next(err)
-    }
-    req.text = string
-    next()
-  })
-});
-
-
-
-app.get(uriFilter, function(req, res){
+/* GET home page. */
+router.get('*', function(req, res){
     console.log('GET -- ' +req.path);
     var filename = uriToFilename(req.path);
     fs.readFile(filename, function(err, data) {
@@ -63,7 +19,7 @@ app.get(uriFilter, function(req, res){
     });
 });
 
-app.put(uriFilter, function(req, res){
+router.put('*', function(req, res){
     console.log('PUT ' +req.path);
     console.log(' text length:' + (req.text ? req.text.length : 'undefined1'))
     var filename = uriToFilename(req.path);
@@ -91,7 +47,7 @@ applySparqlPatch = function() {
   // write me
 }
 
-app.post(uriFilter, function(req, res){
+router.post('*', function(req, res){
     console.log('POST ' +req.path);
     console.log(' text length: ' + (req.text ? req.text.length : 'undefined2'))
     var filename = uriToFilename(req.path);
@@ -212,11 +168,8 @@ app.post(uriFilter, function(req, res){
 
 });
 
-app.patch(uriFilter, function(req, res){
+router.patch('*', function(req, res){
   res.send('Hello World');
 });
 
-var server = app.listen(3000, function() {
-    console.log('Listening on port %d', server.address().port);
-});
-
+module.exports = router;
