+++ package.json
@@ -4,7 +4,7 @@
   "version": "0.0.1",
   "private": true,
   "dependencies": {
-    "express": "3.x",
+    "express": "^3.17.5",
     "mime": ""
   }
 }
